refactor: US1 & US2 refactorized and fixed
Both user stories needed a lot of work because it was practically reinit
the project. I found that the previous implementation was wrong because
it search reservations at a specific time and not the free rooms.

+++ src/main/java/co/edu/udea/salasinfo/persistence/RestrictionDAO.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 
 public interface RestrictionDAO{
-    Restriction findById(Integer id);
+    Restriction findById(Long id);
     List<Restriction> findAll();
     Restriction save(Restriction restriction);
     Restriction findByDescription(String description);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
