+++ src/main/java/co/edu/udea/salasinfo/repository/ImplementRepository.java
@@ -1,9 +1,12 @@
 package co.edu.udea.salasinfo.repository;
 
 import co.edu.udea.salasinfo.model.Implement;
+import co.edu.udea.salasinfo.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ImplementRepository extends JpaRepository<Implement, Integer> {
+public interface ImplementRepository extends JpaRepository<Implement, Long> {
 }
