+++ src/main/java/co/edu/udea/salasinfo/repository/RestrictionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RestrictionRepository extends JpaRepository<Restriction, Integer> {
+public interface RestrictionRepository extends JpaRepository<Restriction, Long> {
     public Optional<Restriction> findByDescription(String description);
 }
